Refactor handlers and their tests
