+++ src/idQuery/queryHandler.test.ts
@@ -1,12 +1,15 @@
-import { _handler as _publicHandler } from './public'
-import { _handler as _idQueryHandler, _splitHandler as _idQuerySplitHandler } from './idQuery';
-import * as dynamodb from './lib/dynamodb'
-import { authenticator, logger, decorate } from './lib/decorators';
+import { _handler as publicHandler } from '../public'
+import { _queryHandler as idQueryHandler } from './queryHandler';
+import * as dynamodb from '../lib/dynamodb'
 
-// TODO: logger、authenticator をテストから分離する
-const publicHandler = decorate(_publicHandler, [logger, authenticator('id-req')]);
-const idQueryHandler = decorate(_idQueryHandler, [logger, authenticator('id-req')]);
-const idQuerySplitHandler = decorate(_idQuerySplitHandler, [logger, authenticator('id-req')]);
+const context = {
+  propIdAuthenticator: { authentication: { plan: 'paid' } },
+  propIdLogger: { background: [] },
+}
+const freeContext = {
+  propIdAuthenticator: { authentication: { plan: 'free' } },
+  propIdLogger: { background: [] },
+}
 
 test('returns 400 when estateId is not available', async () => {
   const event = {
@@ -14,7 +17,7 @@ test('returns 400 when estateId is not available', async () => {
     pathParameters: {}
   }
   // @ts-ignore
-  const lambdaResult = await idQueryHandler(event) as APIGatewayProxyResult
+  const lambdaResult = await idQueryHandler(event, context) as APIGatewayProxyResult
   expect(lambdaResult.statusCode).toBe(400)
 })
 
@@ -26,7 +29,7 @@ test('returns 404 when estateId is not found', async () => {
     }
   }
   // @ts-ignore
-  const lambdaResult = await idQueryHandler(event) as APIGatewayProxyResult
+  const lambdaResult = await idQueryHandler(event, context) as APIGatewayProxyResult
   expect(lambdaResult.statusCode).toBe(404)
 })
 
@@ -38,7 +41,7 @@ test('it works', async () => {
     },
   }
   // @ts-ignore
-  const lambdaResult1 = await publicHandler(event1) as APIGatewayProxyResult
+  const lambdaResult1 = await publicHandler(event1, context) as APIGatewayProxyResult
   const body1 = JSON.parse(lambdaResult1.body)
 
   const event2 = {
@@ -48,7 +51,7 @@ test('it works', async () => {
     }
   }
   // @ts-ignore
-  const lambdaResult2 = await idQueryHandler(event2) as APIGatewayProxyResult
+  const lambdaResult2 = await idQueryHandler(event2, context) as APIGatewayProxyResult
   expect(lambdaResult2.statusCode).toBe(200)
 
   const body2 = JSON.parse(lambdaResult2.body)
@@ -76,7 +79,7 @@ test('should get estate ID without details if authenticated with a free API key'
     }
   }
   // @ts-ignore
-  const lambdaResult1 = await publicHandler(event1) as APIGatewayProxyResult
+  const lambdaResult1 = await publicHandler(event1, freeContext) as APIGatewayProxyResult
   const body1 = JSON.parse(lambdaResult1.body)
 
   const first1 = body1[0]
@@ -106,7 +109,7 @@ test('should get estate ID without details if authenticated with a free API key'
     }
   }
   // @ts-ignore
-  const lambdaResult2 = await idQueryHandler(event2) as APIGatewayProxyResult
+  const lambdaResult2 = await idQueryHandler(event2, freeContext) as APIGatewayProxyResult
   expect(lambdaResult2.statusCode).toBe(200)
 
   const body2 = JSON.parse(lambdaResult2.body)
@@ -133,7 +136,7 @@ test('should get estate ID with details if authenticated with a paid API key', a
     }
   }
   // @ts-ignore
-  const lambdaResult1 = await publicHandler(event1) as APIGatewayProxyResult
+  const lambdaResult1 = await publicHandler(event1, context) as APIGatewayProxyResult
   const body1 = JSON.parse(lambdaResult1.body)
 
   const first1 = body1[0]
@@ -152,7 +155,7 @@ test('should get estate ID with details if authenticated with a paid API key', a
     }
   }
   // @ts-ignore
-  const lambdaResult2 = await idQueryHandler(event2) as APIGatewayProxyResult
+  const lambdaResult2 = await idQueryHandler(event2, context) as APIGatewayProxyResult
   expect(lambdaResult2.statusCode).toBe(200)
 
   const body2 = JSON.parse(lambdaResult2.body)
@@ -183,7 +186,7 @@ test('should not return building name with empty building name parameter', async
     }
   }
   // @ts-ignore
-  const lambdaResult1 = await publicHandler(event1) as APIGatewayProxyResult
+  const lambdaResult1 = await publicHandler(event1, context) as APIGatewayProxyResult
   const body1 = JSON.parse(lambdaResult1.body)
   const first1 = body1[0]
 
@@ -199,7 +202,7 @@ test('should not return building name with empty building name parameter', async
     }
   }
   // @ts-ignore
-  const lambdaResult2 = await idQueryHandler(event2) as APIGatewayProxyResult
+  const lambdaResult2 = await idQueryHandler(event2, context) as APIGatewayProxyResult
   expect(lambdaResult2.statusCode).toBe(200)
   const body2 = JSON.parse(lambdaResult2.body)
 
@@ -220,7 +223,7 @@ test('should not include building name in address2', async () => {
     }
   }
   // @ts-ignore
-  const lambdaResult1 = await publicHandler(event1) as APIGatewayProxyResult
+  const lambdaResult1 = await publicHandler(event1, context) as APIGatewayProxyResult
   const body1 = JSON.parse(lambdaResult1.body)
 
   const event2 = {
@@ -235,7 +238,7 @@ test('should not include building name in address2', async () => {
     }
   }
   // @ts-ignore
-  const lambdaResult2 = await idQueryHandler(event2) as APIGatewayProxyResult
+  const lambdaResult2 = await idQueryHandler(event2, context) as APIGatewayProxyResult
   expect(lambdaResult2.statusCode).toBe(200)
   const body2 = JSON.parse(lambdaResult2.body)
 
@@ -256,7 +259,7 @@ test('should return status parameters', async () => {
     }
   }
   // @ts-ignore
-  const lambdaResult1 = await publicHandler(event1) as APIGatewayProxyResult
+  const lambdaResult1 = await publicHandler(event1, context) as APIGatewayProxyResult
   const body1 = JSON.parse(lambdaResult1.body)
   expect(body1[0].status).toBe('addressPending')
 
@@ -272,101 +275,11 @@ test('should return status parameters', async () => {
     }
   }
   // @ts-ignore
-  const lambdaResult2 = await idQueryHandler(event2) as APIGatewayProxyResult
+  const lambdaResult2 = await idQueryHandler(event2, context) as APIGatewayProxyResult
   expect(lambdaResult2.statusCode).toBe(200)
   const body2 = JSON.parse(lambdaResult2.body)
 
   expect(body2[0].status).toEqual('addressPending')
 })
 
-test('should generate new ID from that of existing.', async () => {
-  // id issue with public handler
-  const event1 = {
-    isDemoMode: true,
-    queryStringParameters: {
-      q: '滋賀県大津市京町４丁目１−１こんにちはビルA棟',
-    },
-  }
-  // @ts-ignore
-  const lambdaResult1 = await publicHandler(event1) as APIGatewayProxyResult
-  expect(lambdaResult1.statusCode).toBe(200)
-  const [idObj1] = JSON.parse(lambdaResult1.body)
 
-  // id split
-  const event2 = {
-    isDemoMode: true,
-    pathParameters: {
-      estateId: idObj1.ID,
-    },
-    queryStringParameters: {
-      lat: '35.1234',
-      lng: '135.1234',
-      building: 'こんにちはビルB棟',
-    }
-  }
-  // @ts-ignore
-  const lambdaResult2 = await idQuerySplitHandler(event2) as APIGatewayProxyResult
-  expect(lambdaResult2.statusCode).toBe(200)
-  const [idObj2] = JSON.parse(lambdaResult2.body)
-  expect(typeof idObj2.ID === 'string').toBeTruthy()
-  expect(idObj2.ID).not.toBe(idObj1.ID)
-  expect(idObj2.status).toEqual('addressPending')
-
-  // id query with public handler
-  const event3 = {
-    isDemoMode: true,
-    queryStringParameters: {
-      q: '滋賀県大津市京町４丁目１−１',
-    },
-  }
-  // @ts-ignore
-  const lambdaResult3 = await publicHandler(event3) as APIGatewayProxyResult
-  expect(lambdaResult3.statusCode).toBe(200)
-  const idObjects3 = JSON.parse(lambdaResult3.body)
-  expect(idObjects3).toHaveLength(2)
-  if(idObjects3[0].ID === idObj1.ID) {
-    expect(idObjects3[0]).toMatchObject(idObj1)
-    expect(idObjects3[1]).toMatchObject(idObj2)
-  } else {
-    expect(idObjects3[0]).toMatchObject(idObj2)
-    expect(idObjects3[1]).toMatchObject(idObj1)
-  }
-
-  // id query with query handler
-  const [event4, event5] = idObjects3.map((idObj: any) => ({
-    isDemoMode: true,
-    pathParameters: {
-      estateId: idObj.ID,
-    },
-  }))
-
-  // @ts-ignore
-  const [lambdaResult4, lambdaResult5] = await Promise.all([
-    // @ts-ignore
-    idQueryHandler(event4),
-    // @ts-ignore
-    idQueryHandler(event5),
-  ])
-  // @ts-ignore
-  expect(lambdaResult4.statusCode).toBe(200)
-  // @ts-ignore
-  expect(lambdaResult5.statusCode).toBe(200)
-  // @ts-ignore
-  const [idObj4] = JSON.parse(lambdaResult4.body)
-  // @ts-ignore
-  const [idObj5] = JSON.parse(lambdaResult5.body)
-
-  if(idObj4.ID === idObj1.ID) {
-    expect(idObj4).toMatchObject(idObj1)
-  } else {
-    expect(idObj5).toMatchObject(idObj1)
-  }
-
-  if(idObj5.ID === idObjects3[0].ID) {
-    expect(idObj5).toMatchObject(idObjects3[0])
-    expect(idObj4).toMatchObject(idObjects3[1])
-  } else {
-    expect(idObj5).toMatchObject(idObjects3[1])
-    expect(idObj4).toMatchObject(idObjects3[0])
-  }
-})
