+++ src/public.test.ts
@@ -3,11 +3,9 @@ import { authenticator, decorate, logger } from './lib/decorators'
 import * as dynamodb from './lib/dynamodb'
 import { _getServiceUsageQuotaItem, _updateServiceUsageQuota } from './lib/dynamodb_test_helpers.test'
 import { _handler } from './public'
-import { _splitHandler } from './idQuery'
 
 // TODO: logger、authenticator をテストから分離する
 const handler = decorate(_handler, [logger, authenticator('id-req')]);
-const idQuerySplitHandler = decorate(_splitHandler, [logger, authenticator('id-req')]);
 
 test('should specify the ZOOM environmental variable.', () => {
   const ZOOM = parseInt(process.env.ZOOM, 10)
